Inactive users no longer allowed to move in.
Added story spec to test inactive user login case.
Added active_for_authentication? test to user model to ensure user is
active.

Note that if an inactive user attempts to log in, this will fail
silently, which could cause confustion. Ideally, I would like to see an
error message of some kind indicating why the login attempt was
rejected. As it is, it is consistent with other login failures (i.e. all
login failures give no indication of what the reaon for the rejection
is). So, it's sufficient as is, but I would prefer to reconsider at some
point having a flash message in this instance (and others).

Also, note that due to an apparent bug in master, this branch currently
has a failure in the spec, albeit one which at least seems to be
completely independent from the work done in this branch.

[Finishes #55200014]