Select box works well and CSS styling is good. One test still failing.
Need some guidance on how to make Capybara select an option when the DOM
just has a <ul>.

Also need to move all of the styling from tweaks in the
jquery.selectBoxIt.css file to be overrides in the layout.sass file.

Finally, some CSS cleanups (like making the arrow in the select button
match) would be good, but not necessary.

[#55200222]