Cleaned up database queries, implemented cacheing.
DB queries worked, but were going to scale poorly. Now they are more
correct. Implemented one layer of cache, keyed to user and most recent
work unit start time, which should obviate the need for expiring the
cache.

More work could be done to have more granular cache options, but this is
sufficient for the moment.

[Finishes #48441443]