Removed "_source" check in git_commits_for
The previous code for github_url_source doesn't have a good replacement now that
the github_url attrinute is gone, but it's not clear to me why the _source call
is needed at the moment. Moreover, activities are only associated with projects
not repositories, so they only need to be pulled from the project for now.
Resolves all of the self_and_descendants errors in rspec.