Modify github update for multiple repos
If more than one repo has the same url, the commits in the push event
will be created for all associated projects.

The project method and instance variable in GithubPull and GithubUpdate
has been changed to "projects," which returns an array of projects.

The uniqueness check in GithubCommit was changed to account for project
and id. This implementation is likely to be resource intensive, and should
be changed.

GithubControllerSpec was updated to match the new information in the
push event payload stub.