Move id from properties to source_id column
Moved the id from the properties hstore to the source_id column
in the activities table. There are three migrations: one to create
the column, one to move the information into the column and remove
the id key from properties, and one to create a unique index for
project, source, and source_id. Subsequent changes needed to models
and controllers involving activities to use the appropriate attribute.

The change introduced a problem with FactoryGirl creating activities,
so they have been replaced with before methods.