Associate github commits with work_unit
Github commits now associate with a work_unit if there is one that
includes the timestamp for the same project and user. There is a
"find_work_unit" method in PivotalActivity that is blank for now.

Added byebug to test environment.

Although there was Activity model included the association
"belongs_to :work_unit", the activities table had no foreign key for
work units. A migration added the column and also added the inverse
association in the WorkUnit model.

I had trouble with the PivotalPullController spec, and I am not
confident that the current specs adequately test the Pivotal
pull/update features.