Replace notes field with notes method
The work-units field notes has been replaced with a method notes that
concatenates all of the annotations for a work unit into a single
string. Also added a new factory for work_unit_with_annotation that
can be used to create a work unit that meets the previous requirements
for having the notes field present.